pyproject.toml
[project]
name = "chess-engine-2"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.8.0",
    "python-dotenv>=1.0.1",
    "numpy>=2.3.2",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[project.scripts]
train = "chess_engine_2.train:cli"
